Add example files to e2e folder; some refactoring
