--- src/vscode-bicep/src/test/utils/index.ts
@@ -1,5 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-export * from "./assert";
-export * from "./string";
-export * from "./fs";
